feature/DIPA-234 links structure (#144)
* feat: DIPA-234 created menu for home with "project tour" and "useful links" item
restructured gantt and overview module to suit the look and feel

* fix: DIPA-234 fixed HomeComponent route and redirect '' to home/tour

* fix: DIPA-234 fixed GanttComponent route

* fix: removed console.log

* fix: reformated code with prettier
