+++ frontend/projects/dipa-frontend/e2e/tsconfig.json
@@ -5,10 +5,6 @@
     "outDir": "../../../out-tsc/e2e",
     "module": "commonjs",
     "target": "es2018",
-    "types": [
-      "jasmine",
-      "jasminewd2",
-      "node"
-    ]
+    "types": ["jasmine", "jasminewd2", "node"]
   }
 }
