+++ frontend/projects/dipa-frontend/src/app/app.component.ts
@@ -7,6 +7,6 @@ import { GanttMenuComponent } from './menus/gantt-menu/gantt-menu.component';
   styleUrls: ['./app.component.scss'],
 })
 export class AppComponent {
-  title = 'dipa-frontend';
   @ViewChild(GanttMenuComponent) ganttMenuComponent: GanttMenuComponent;
+  title = 'dipa-frontend';
 }
