+++ frontend/projects/dipa-frontend/src/app/menus/gantt-menu/gantt-menu.component.html
@@ -1,5 +1,5 @@
 <mat-menu menu="matMenu">
   <div *ngFor="let timeline of timelineData">
-    <button mat-menu-item [routerLink]="['/gantt/', timeline.id]">{{ timeline.name }}</button>
+    <button mat-menu-item [routerLink]="['/gantt/', timeline.id, 'timeline']">{{ timeline.name }}</button>
   </div>
 </mat-menu>
