+++ frontend/projects/dipa-frontend/src/app/modules/gantt/gantt.component.spec.ts
@@ -1,17 +1,14 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { GanttComponent } from './gantt.component';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { RouterTestingModule } from '@angular/router/testing';
-import { MatMenuModule } from '@angular/material/menu';
-import { MaterialModule } from '../../material/material.module';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { FormsModule } from '@angular/forms';
+import { MatSidenav, MatSidenavContainer, MatSidenavContent } from '@angular/material/sidenav';
 import { SidenavComponent } from './sidenav/sidenav.component';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { ActivatedRoute } from '@angular/router';
+import { of } from 'rxjs';
 import { MatNavList } from '@angular/material/list';
-import { MatIcon } from '@angular/material/icon';
-import { NavMenuListItemComponent } from '../../shared/nav-menu-list-item/nav-menu-list-item.component';
-import { NavService } from '../../nav.service';
+import { RouterTestingModule } from '@angular/router/testing';
 
 describe('GanttComponent', () => {
   let component: GanttComponent;
@@ -19,16 +16,18 @@ describe('GanttComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [GanttComponent, SidenavComponent, MatNavList, MatIcon, NavMenuListItemComponent],
-      imports: [
-        RouterTestingModule,
-        HttpClientTestingModule,
-        MatMenuModule,
-        MaterialModule,
-        BrowserAnimationsModule,
-        FormsModule,
+      declarations: [GanttComponent, MatSidenav, MatSidenavContainer, SidenavComponent, MatSidenavContent, MatNavList],
+      imports: [RouterTestingModule, BrowserAnimationsModule, HttpClientTestingModule],
+      providers: [
+        {
+          provide: ActivatedRoute,
+          useValue: {
+            params: of({
+              id: '1',
+            }),
+          },
+        },
       ],
-      providers: [NavService],
     }).compileComponents();
   });
