+++ frontend/projects/dipa-frontend/src/app/modules/gantt/templates-view/templates-view.component.html
@@ -1,49 +1,44 @@
 <div id="gantt" class="gantt" #gantt>
-  <mat-sidenav-container autosize>
-    <mat-sidenav mode="side" opened="true" class="sidenav">
-      <app-sidenav></app-sidenav>
-    </mat-sidenav>
-    <mat-sidenav-content class="content-container">
-      <div class="content-container">
-        <div class="row justify-content-between">
-          <div class="col">
-            <div class="row justify-content-start">
-              <div class="operationType-container">
-                <p class="label">Vorhabensart</p>
-                <p class="name">{{ selectedOperationTypeName }}</p>
-              </div>
-              <div *ngIf="projectApproachesList" class="col">
-                <mat-form-field appearance="fill" floatLabel="always">
-                  <mat-label>Vorgehensweise</mat-label>
-                  <mat-select [(value)]="selectedProjectApproachId" (selectionChange)="changeProjectApproach($event)">
-                    <mat-option *ngFor="let projectApproach of filterProjectApproaches()" [value]="projectApproach.id">
-                      {{ projectApproach.name }}
-                    </mat-option>
-                  </mat-select>
-                </mat-form-field>
-              </div>
-            </div>
+  <div class="content-container" style="padding-right: 15px">
+    <div class="row justify-content-between">
+      <div class="col">
+        <div class="row justify-content-start">
+          <div class="operationType-container">
+            <p class="label">Vorhabensart</p>
+            <p class="name">{{ selectedOperationTypeName }}</p>
           </div>
-          <div class="col">
-            <div class="row justify-content-end">
-              <mat-button-toggle-group
-                name="chartMetric"
-                #group="matButtonToggleGroup"
-                multiple
-                aria-label="Chart Metric"
-                (change)="changeViewType($event)"
-              >
-                <mat-button-toggle class="md-button" value="YEARS">Jahre</mat-button-toggle>
-                <mat-button-toggle class="md-button" value="MONTHS">Monate</mat-button-toggle>
-                <mat-button-toggle class="md-button" value="WEEKS">Wochen</mat-button-toggle>
-                <mat-button-toggle class="md-button" value="DAYS">Tage</mat-button-toggle>
-              </mat-button-toggle-group>
-            </div>
+          <div *ngIf="projectApproachesList" class="col">
+            <mat-form-field appearance="fill" floatLabel="always">
+              <mat-label>Vorgehensweise</mat-label>
+              <mat-select [(value)]="selectedProjectApproachId" (selectionChange)="changeProjectApproach($event)">
+                <mat-option *ngFor="let projectApproach of filterProjectApproaches()" [value]="projectApproach.id">
+                  {{ projectApproach.name }}
+                </mat-option>
+              </mat-select>
+            </mat-form-field>
           </div>
         </div>
       </div>
-      <div *ngIf="timelineData" class="row chart-container">
-        <div class="col-2 template-controller-container">
+      <div class="col">
+        <div class="row justify-content-end">
+          <mat-button-toggle-group
+            name="chartMetric"
+            #group="matButtonToggleGroup"
+            multiple
+            aria-label="Chart Metric"
+            (change)="changeViewType($event)"
+          >
+            <mat-button-toggle class="md-button" value="YEARS">Jahre</mat-button-toggle>
+            <mat-button-toggle class="md-button" value="MONTHS">Monate</mat-button-toggle>
+            <mat-button-toggle class="md-button" value="WEEKS">Wochen</mat-button-toggle>
+            <mat-button-toggle class="md-button" value="DAYS">Tage</mat-button-toggle>
+          </mat-button-toggle-group>
+        </div>
+      </div>
+    </div>
+    <div *ngIf="timelineData" class="row chart-container">
+      <div class="col template-controller-container">
+        <div>
           <div class="template-controller">
             <h2>Aktuell</h2>
           </div>
@@ -78,20 +73,18 @@
             </div>
           </div>
         </div>
-        <div class="col-10">
-          <div class="">
-            <ng-container *ngIf="vm$ | async as vm">
-              <app-templates
-                #templateChart
-                [timelineData]="vm.selectedTimeline"
-                [templateData]="vm.selectedTemplates"
-                [allTemplates]="vm.templatesData"
-                id="template{{ vm.selectedTimeline.id }}"
-              ></app-templates>
-            </ng-container>
-          </div>
+        <div class="full-width">
+          <ng-container *ngIf="vm$ | async as vm">
+            <app-templates
+              #templateChart
+              [timelineData]="vm.selectedTimeline"
+              [templateData]="vm.selectedTemplates"
+              [allTemplates]="vm.templatesData"
+              id="template{{ vm.selectedTimeline.id }}"
+            ></app-templates>
+          </ng-container>
         </div>
       </div>
-    </mat-sidenav-content>
-  </mat-sidenav-container>
+    </div>
+  </div>
 </div>
