+++ frontend/projects/dipa-frontend/src/app/modules/gantt/timeline/timeline.component.html
@@ -1,0 +1,69 @@
+<div id="gantt" class="gantt" #gantt>
+  <div class="content-container" style="padding-right: 15px">
+    <div class="row justify-content-between">
+      <div class="col-6">
+        <div class="row justify-content-start">
+          <div class="operationType-container">
+            <p class="label">Vorhabensart</p>
+            <p class="name">{{ selectedOperationTypeName }}</p>
+          </div>
+          <div *ngIf="timelineData" class="col">
+            <mat-form-field appearance="fill" floatLabel="always">
+              <mat-label>Vorgehensweise</mat-label>
+              <mat-select [(value)]="selectedProjectApproachId" (selectionChange)="changeProjectApproach($event)">
+                <mat-option *ngFor="let projectApproach of filterProjectApproaches()" [value]="projectApproach.id">
+                  {{ projectApproach.name }}
+                </mat-option>
+              </mat-select>
+            </mat-form-field>
+          </div>
+        </div>
+      </div>
+      <div class="col-6">
+        <div class="row justify-content-end">
+          <mat-button-toggle-group
+            name="chartMetric"
+            #group="matButtonToggleGroup"
+            multiple
+            aria-label="Chart Metric"
+            [(ngModel)]="viewTypeSelected"
+            (change)="changeViewType($event)"
+          >
+            <mat-button-toggle class="md-button" value="YEARS">Jahre</mat-button-toggle>
+            <mat-button-toggle class="md-button" value="MONTHS">Monate</mat-button-toggle>
+            <mat-button-toggle class="md-button" value="WEEKS">Wochen</mat-button-toggle>
+            <mat-button-toggle class="md-button" value="DAYS">Tage</mat-button-toggle>
+          </mat-button-toggle-group>
+        </div>
+      </div>
+    </div>
+    <div class="row">
+      <div class="col">
+        <button
+          (click)="getIcsCalendarFile()"
+          mat-raised-button
+          title="Exportiere Meilensteine im iCal Format"
+          aria-label="Exportiere Meilensteine im iCal Format"
+          style="margin-bottom: 3px"
+        >
+          <mat-icon>event_note</mat-icon>
+          Export
+        </button>
+      </div>
+    </div>
+    <div class="row">
+      <div class="col">
+        <ng-container *ngIf="vm$ | async as vm">
+          <app-chart
+            #ganttChart
+            [timelineData]="vm.selectedTimeline"
+            [milestoneData]="vm.milestoneData"
+            [taskData]="vm.taskData"
+            [incrementsData]="vm.incrementsData"
+            id="gantt{{ vm.selectedTimeline.id }}"
+          ></app-chart>
+        </ng-container>
+      </div>
+    </div>
+  </div>
+</div>
