+++ frontend/projects/dipa-frontend/src/app/modules/gantt/timeline/timeline.component.scss
@@ -44,34 +44,3 @@ mat-sidenav-content {
   text-overflow: ellipsis;
   overflow: hidden;
 }
-
-.template-controller {
-  text-align: center;
-  justify-content: center;
-  margin: auto;
-  padding-top: 140px;
-}
-
-.template-controller-container {
-  display: inline-flex;
-}
-
-.logo {
-  width: 100px;
-}
-
-button {
-  margin: 5px;
-}
-
-.button-template {
-  text-align: center;
-}
-
-.text-template {
-  margin: 0;
-}
-
-.full-width {
-  width: 100%;
-}
