+++ frontend/projects/dipa-frontend/src/app/modules/gantt/toolkit/toolkit.component.html
@@ -1,12 +1,1 @@
-<div id="gantt" class="gantt" #gantt>
-  <mat-sidenav-container autosize>
-    <mat-sidenav mode="side" opened="true" class="sidenav">
-      <app-sidenav></app-sidenav>
-    </mat-sidenav>
-    <mat-sidenav-content class="content-container">
-      <div class="content-container">
-        <p>toolkit works!</p>
-      </div>
-    </mat-sidenav-content>
-  </mat-sidenav-container>
-</div>
+<p>toolkit works!</p>
