+++ frontend/projects/dipa-frontend/src/app/modules/home/home-routing.module.ts
@@ -1,9 +1,20 @@
 import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
+import { RouterModule, Routes } from '@angular/router';
 
 import { HomeComponent } from './home.component';
+import { ProjectTourComponent } from './project-tour/project-tour.component';
+import { LinksComponent } from './links/links.component';
 
-const routes: Routes = [{ path: '', component: HomeComponent }];
+const routes: Routes = [
+  {
+    path: '',
+    component: HomeComponent,
+    children: [
+      { path: 'tour', component: ProjectTourComponent },
+      { path: 'links', component: LinksComponent },
+    ],
+  },
+];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
