+++ frontend/projects/dipa-frontend/src/app/modules/home/home.component.spec.ts
@@ -4,6 +4,11 @@ import { HomeComponent } from './home.component';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { MatNavList } from '@angular/material/list';
 import { MatIcon } from '@angular/material/icon';
+import { MatSidenav, MatSidenavContainer, MatSidenavContent } from '@angular/material/sidenav';
+import { RouterTestingModule } from '@angular/router/testing';
+import { NavMenuListItemComponent } from '../../shared/nav-menu-list-item/nav-menu-list-item.component';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { NavService } from '../../nav.service';
 
 describe('HomeComponent', () => {
   let component: HomeComponent;
@@ -11,8 +16,18 @@ describe('HomeComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [HomeComponent, MatNavList, MatIcon],
-      imports: [HttpClientTestingModule],
+      declarations: [
+        HomeComponent,
+        MatNavList,
+        MatIcon,
+        MatSidenavContainer,
+        MatSidenav,
+        MatSidenavContent,
+        NavMenuListItemComponent,
+        MatNavList,
+      ],
+      imports: [BrowserAnimationsModule, RouterTestingModule, HttpClientTestingModule],
+      providers: [NavService],
     }).compileComponents();
   });
