+++ frontend/projects/dipa-frontend/src/app/modules/home/home.component.ts
@@ -1,35 +1,32 @@
-import { Component, OnDestroy, OnInit } from '@angular/core';
-import { ExternalLinksService } from 'dipa-api-client';
+import { Component, OnInit } from '@angular/core';
+import { NavItem } from '../../nav-item';
 
 @Component({
   selector: 'app-home',
   templateUrl: './home.component.html',
   styleUrls: ['./home.component.scss'],
 })
-export class HomeComponent implements OnInit, OnDestroy {
-  externalLinksSubscription;
+export class HomeComponent implements OnInit {
+  navMenuItems: NavItem[] = [];
 
-  externalLinkGroups;
-
-  constructor(private externalLinksService: ExternalLinksService) {}
+  constructor() {}
 
   ngOnInit(): void {
-    this.externalLinksSubscription = this.externalLinksService.getExternalLinks().subscribe((data) => {
-      this.externalLinkGroups = Object.create(null);
-
-      data.forEach((link) => {
-        this.externalLinkGroups[link.category] = this.externalLinkGroups[link.category] || [];
-        this.externalLinkGroups[link.category].push(link);
-      });
-    });
+    this.setSideNavMenu();
   }
 
-  ngOnDestroy(): void {
-    this.externalLinksSubscription?.unsubscribe();
-  }
-
-  // no ordering
-  returnZero(): number {
-    return 0;
+  setSideNavMenu(): void {
+    this.navMenuItems = [
+      {
+        name: 'Eine Reise durchs Projekt',
+        icon: 'explore',
+        route: 'home/tour',
+      },
+      {
+        name: 'Nützliche Links',
+        icon: 'bookmarks',
+        route: 'home/links',
+      },
+    ];
   }
 }
