+++ frontend/projects/dipa-frontend/src/app/modules/home/project-tour/project-tour.component.html
@@ -1,0 +1,1 @@
+<p>project-tour works!</p>
