+++ frontend/projects/dipa-frontend/src/app/modules/home/project-tour/project-tour.component.ts
@@ -1,0 +1,12 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-project-tour',
+  templateUrl: './project-tour.component.html',
+  styleUrls: ['./project-tour.component.scss'],
+})
+export class ProjectTourComponent implements OnInit {
+  constructor() {}
+
+  ngOnInit(): void {}
+}
