+++ frontend/projects/dipa-frontend/src/app/modules/overview/overview.component.scss
@@ -1,17 +1,6 @@
-
-h1 {
-    text-align: center;
-}
-
-h2 {
-    color: #3f51b5;
-}
-
-.row {
-    align-items: center;
-    display: flex;
-}
-
-div.gantt {
-  margin-bottom: 1.34375em;
+.mat-sidenav-container {
+  width: 100%;
+  height: calc(100vh - 84px);
+  background-color: transparent;
+  margin-top: 15px;
 }
