+++ frontend/projects/dipa-frontend/src/app/modules/overview/overview.component.spec.ts
@@ -3,6 +3,12 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { OverviewComponent } from './overview.component';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { MatSidenavModule } from '@angular/material/sidenav';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { NavMenuListItemComponent } from '../../shared/nav-menu-list-item/nav-menu-list-item.component';
+import { NavService } from '../../nav.service';
+import { RouterTestingModule } from '@angular/router/testing';
+import { MatIcon } from '@angular/material/icon';
+import { MatNavList } from '@angular/material/list';
 
 describe('OverviewComponent', () => {
   let component: OverviewComponent;
@@ -10,8 +16,9 @@ describe('OverviewComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [OverviewComponent],
-      imports: [HttpClientTestingModule, MatSidenavModule],
+      declarations: [OverviewComponent, NavMenuListItemComponent, MatIcon, MatNavList],
+      imports: [HttpClientTestingModule, RouterTestingModule, MatSidenavModule, BrowserAnimationsModule],
+      providers: [NavService],
     }).compileComponents();
   });
