+++ frontend/projects/dipa-frontend/src/app/modules/overview/projects/projects.component.html
@@ -1,0 +1,62 @@
+<div class="">
+  <div class="row justify-content-center">
+    <mat-spinner *ngIf="loading"></mat-spinner>
+  </div>
+  <ng-container *ngFor="let vm of vmAll$ | async">
+    <div class="content-container" style="padding-right: 15px">
+      <div class="row">
+        <div class="col col-md-auto">
+          <div class="mat-card-title">
+            {{ vm.timeline.name }}
+          </div>
+          <div class="mat-card-subtitle">
+            {{ vm.timeline.projectType }}
+          </div>
+        </div>
+        <div *ngIf="timelineData" class="col col-md-auto">
+          <mat-form-field appearance="fill" floatLabel="always" id="operationType-{{ vm.timeline.id }}">
+            <mat-label>Vorhabensart</mat-label>
+            <mat-select
+              [(value)]="vm.timeline.operationTypeId"
+              (selectionChange)="changeOperationType($event, vm.timeline.id)"
+            >
+              <mat-option *ngFor="let operationType of operationTypesList" [value]="operationType.id">
+                {{ operationType.name }}
+              </mat-option>
+            </mat-select>
+          </mat-form-field>
+        </div>
+        <div *ngIf="timelineData" class="col col-md-auto">
+          <mat-form-field appearance="fill" floatLabel="always" id="projectApproach-{{ vm.timeline.id }}">
+            <mat-label>Vorgehensweise</mat-label>
+            <mat-select
+              [(value)]="vm.timeline.projectApproachId"
+              (selectionChange)="changeProjectApproach($event, vm.timeline.id)"
+            >
+              <mat-option
+                *ngFor="let projectApproach of filterProjectApproaches(vm.timeline.operationTypeId)"
+                [value]="projectApproach.id"
+              >
+                {{ projectApproach.name }}
+              </mat-option>
+            </mat-select>
+          </mat-form-field>
+        </div>
+      </div>
+      <div class="row">
+        <div class="col">
+          <div id="gantt{{ vm.timeline.id }}" class="gantt">
+            <app-chart
+              #charts
+              [timelineData]="vm.timeline"
+              [milestoneData]="vm.milestoneData"
+              [taskData]="vm.taskData"
+              [incrementsData]="vm.incrementsData"
+              id="overview{{ vm.timeline.id }}"
+            ></app-chart>
+          </div>
+        </div>
+      </div>
+    </div>
+  </ng-container>
+</div>
