+++ frontend/projects/dipa-frontend/src/app/modules/overview/projects/projects.component.scss
@@ -1,0 +1,22 @@
+
+h1 {
+  text-align: center;
+}
+
+h2 {
+  color: #3f51b5;
+}
+
+.row {
+  align-items: center;
+  display: flex;
+}
+
+div.gantt {
+  margin-bottom: 1.34375em;
+}
+
+.content-container {
+  margin: 10px;
+  padding-left: 10px;
+}
